+++ types/protocol-mapping.d.ts
@@ -356,10 +356,10 @@ export namespace ProtocolMapping {
         'Media.playerErrorsRaised': [Protocol.Media.PlayerErrorsRaisedEvent];
         /**
          * Called whenever a player is created, or when a new agent joins and receives
-         * a list of active players. If an agent is restored, it will receive the full
-         * list of player ids and all events again.
+         * a list of active players. If an agent is restored, it will receive one
+         * event for each active player.
          */
-        'Media.playersCreated': [Protocol.Media.PlayersCreatedEvent];
+        'Media.playerCreated': [Protocol.Media.PlayerCreatedEvent];
         /**
          * Fired when data chunk was received over the network.
          */
@@ -4055,6 +4055,15 @@ export namespace ProtocolMapping {
         'Memory.getSamplingProfile': {
             paramsType: [];
             returnType: Protocol.Memory.GetSamplingProfileResponse;
+        };
+        /**
+         * Returns enum representing if IP Proxy of requests is available
+         * or reason it is not active.
+         * @experimental
+         */
+        'Network.getIPProtectionProxyStatus': {
+            paramsType: [];
+            returnType: Protocol.Network.GetIPProtectionProxyStatusResponse;
         };
         /**
          * Sets a list of content encodings that will be accepted. Empty list means no encoding is accepted.