+++ types/protocol-proxy-api.d.ts
@@ -2875,10 +2875,10 @@ export namespace ProtocolProxyApi {
 
         /**
          * Called whenever a player is created, or when a new agent joins and receives
-         * a list of active players. If an agent is restored, it will receive the full
-         * list of player ids and all events again.
+         * a list of active players. If an agent is restored, it will receive one
+         * event for each active player.
          */
-        on(event: 'playersCreated', listener: (params: Protocol.Media.PlayersCreatedEvent) => void): void;
+        on(event: 'playerCreated', listener: (params: Protocol.Media.PlayerCreatedEvent) => void): void;
 
     }
 
@@ -2945,6 +2945,13 @@ export namespace ProtocolProxyApi {
     }
 
     export interface NetworkApi {
+        /**
+         * Returns enum representing if IP Proxy of requests is available
+         * or reason it is not active.
+         * @experimental
+         */
+        getIPProtectionProxyStatus(): Promise<Protocol.Network.GetIPProtectionProxyStatusResponse>;
+
         /**
          * Sets a list of content encodings that will be accepted. Empty list means no encoding is accepted.
          * @experimental